requirements.txt
# NASDAQ Volatility Scanner - Dependencies
streamlit>=1.30.0
alpaca-py>=0.21.0
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
ta>=0.10.2
requests>=2.28.0
python-dotenv>=1.0.0
schedule>=1.2.0
colorlog>=6.7.0
